+++ posthog-react-native/src/posthog-rn.ts
@@ -212,10 +212,25 @@ export class PostHog extends PostHogCore {
     )
   }
 
+  _isEnableSessionReplay(): boolean {
+    return !this.isDisabled && (this._enableSessionReplay ?? false)
+  }
+
+  _resetSessionId(
+    reactNativeSessionReplay: typeof OptionalReactNativeSessionReplay | undefined,
+    sessionId: string
+  ): void {
+    // _resetSessionId is only called if reactNativeSessionReplay not undefined, but the linter wasn't happy
+    if (reactNativeSessionReplay) {
+      reactNativeSessionReplay.endSession()
+      reactNativeSessionReplay.startSession(sessionId)
+    }
+  }
+
   getSessionId(): string {
     const sessionId = super.getSessionId()
 
-    if (!this._enableSessionReplay) {
+    if (!this._isEnableSessionReplay()) {
       return sessionId
     }
 
@@ -223,8 +238,7 @@ export class PostHog extends PostHogCore {
     if (sessionId.length > 0 && this._currentSessionId && sessionId !== this._currentSessionId) {
       if (OptionalReactNativeSessionReplay) {
         try {
-          OptionalReactNativeSessionReplay.endSession()
-          OptionalReactNativeSessionReplay.startSession(sessionId)
+          this._resetSessionId(OptionalReactNativeSessionReplay, sessionId)
           this.logMsgIfDebug(() => console.info('PostHog Debug', `Session replay started with sessionId ${sessionId}.`))
         } catch (e) {
           this.logMsgIfDebug(() =>
@@ -233,6 +247,11 @@ export class PostHog extends PostHogCore {
         }
       }
       this._currentSessionId = sessionId
+    } else {
+      console.log(
+        'PostHog Debug',
+        `Session replay session id not rotated, sessionId ${sessionId} and currentSessionId ${this._currentSessionId}.`
+      )
     }
 
     return sessionId
@@ -240,7 +259,7 @@ export class PostHog extends PostHogCore {
 
   resetSessionId(): void {
     super.resetSessionId()
-    if (this._enableSessionReplay && OptionalReactNativeSessionReplay) {
+    if (this._isEnableSessionReplay() && OptionalReactNativeSessionReplay) {
       try {
         OptionalReactNativeSessionReplay.endSession()
         this.logMsgIfDebug(() => console.info('PostHog Debug', `Session replay ended.`))
@@ -254,7 +273,7 @@ export class PostHog extends PostHogCore {
     const previousDistinctId = this.getDistinctId()
     super.identify(distinctId, properties, options)
 
-    if (this._enableSessionReplay && OptionalReactNativeSessionReplay) {
+    if (this._isEnableSessionReplay() && OptionalReactNativeSessionReplay) {
       try {
         distinctId = distinctId || previousDistinctId
         OptionalReactNativeSessionReplay.identify(distinctId, this.getAnonymousId())
@@ -273,18 +292,27 @@ export class PostHog extends PostHogCore {
 
   private async startSessionReplay(options?: PostHogOptions): Promise<void> {
     this._enableSessionReplay = options?.enableSessionReplay
-    if (!this._enableSessionReplay) {
+    if (!this._isEnableSessionReplay()) {
       this.logMsgIfDebug(() => console.info('PostHog Debug', 'Session replay is not enabled.'))
       return
     }
 
-    const sdkReplayConfig = options?.sessionReplayConfig ?? {
-      maskAllTextInputs: true,
-      maskAllImages: true,
-      captureLog: true,
-      captureNetworkTelemetry: true,
-      iOSdebouncerDelayMs: 1000,
-      androidDebouncerDelayMs: 500,
+    const {
+      maskAllTextInputs = true,
+      maskAllImages = true,
+      captureLog = true,
+      captureNetworkTelemetry = true,
+      iOSdebouncerDelayMs = 1000,
+      androidDebouncerDelayMs = 1000,
+    } = options?.sessionReplayConfig ?? {}
+
+    const sdkReplayConfig = {
+      maskAllTextInputs,
+      maskAllImages,
+      captureLog,
+      captureNetworkTelemetry,
+      iOSdebouncerDelayMs,
+      androidDebouncerDelayMs,
     }
 
     this.logMsgIfDebug(() =>
@@ -316,6 +344,7 @@ export class PostHog extends PostHogCore {
           distinctId: this.getDistinctId(),
           anonymousId: this.getAnonymousId(),
           sdkVersion: this.getLibraryVersion(),
+          flushAt: this.flushAt,
         }
 
         this.logMsgIfDebug(() =>
@@ -329,7 +358,11 @@ export class PostHog extends PostHogCore {
               console.info('PostHog Debug', `Session replay started with sessionId ${sessionId}.`)
             )
           } else {
-            this.logMsgIfDebug(() => console.log('PostHog Debug', `Session replay already started.`))
+            // if somehow the SDK is already enabled with a different sessionId, we reset it
+            this._resetSessionId(OptionalReactNativeSessionReplay, sessionId)
+            this.logMsgIfDebug(() =>
+              console.log('PostHog Debug', `Session replay already started with sessionId ${sessionId}.`)
+            )
           }
         } catch (e) {
           this.logMsgIfDebug(() => console.error('PostHog Debug', `Session replay failed to start: ${e}.`))